backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
httpx
pydantic
python-dotenv
bandit
radon
packaging
